+++ module/actor/character-sheet-gen4.js
@@ -2,13 +2,13 @@
  * Extend the basic ActorSheet with some very simple modifications
  * @extends {ActorSheet}
  */
-export class PTUCharacterSheet extends ActorSheet {
+export class PTUGen4CharacterSheet extends ActorSheet {
 
   /** @override */
   static get defaultOptions() {
     return mergeObject(super.defaultOptions, {
-      classes: ["ptu", "sheet", "actor"],
-      template: "systems/ptu/templates/actor/character-sheet.hbs",
+      classes: ["ptu", "sheet", "actor", "gen4"],
+      template: "systems/ptu/templates/actor/character-sheet-gen4.hbs",
       width: 675,
       height: 675,
       tabs: [{ navSelector: ".sheet-tabs", contentSelector: ".sheet-body", initial: "stats" }]